refactor(transactions): extract top-level feature components
